moved instrumentation src into its own directory
