+++ core/instrumentation/st-data-collection.c
