+++ core/instrumentation/st-stats-buffer.c
