+++ core/ross.h
@@ -216,9 +216,9 @@ typedef uint64_t tw_lpid;
 
 #include "ross-inline.h"
 
-#include "st-stats-buffer.h"
-#include "st-data-collection.h"
-#include "st-event-collection.h"
+#include "instrumentation/st-stats-buffer.h"
+#include "instrumentation/st-data-collection.h"
+#include "instrumentation/st-event-collection.h"
 
 #ifdef __cplusplus
 }
